slurm-runner/requirements.txt
black
isort
matplotlib
optuna
pandas
plotly
pytorch-ignite
pytorch-metric-learning
PyYAML
record-keeper==0.9.31.dev3
scikit-learn
scipy
submitit
tensorboard
timm==0.4.9
torchmetrics==0.3.2
tqdm
